Reviewed-on: http://192.168.1.126:3000/BANNARI/BSOFT/pulls/601

+++ BackgroundService/BackgroundService.Application/Notification/NotificationGroup/Queries/GetAllNotificationGroup/NotificationGroupDto.cs
@@ -9,7 +9,7 @@ namespace BackgroundService.Application.Notification.NotificationGroup.Queries.G
     {
         public int Id { get; set; }
         public string? GroupName { get; set; }
-        public byte IsActive { get; set; }
+        public int IsActive { get; set; }
         public int CreatedBy { get; set; }
         public DateTimeOffset CreatedDate { get; set; }
         public string? CreatedByName { get; set; }
